+++ src/main/java/subway/controller/LineController.java
@@ -1,8 +1,14 @@
-package subway.line;
+package subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.station.StationDao;
+import subway.domain.Line;
+import subway.domain.Section;
+import subway.response.*;
+import subway.request.LineRequest;
+import subway.request.SectionRequest;
+import subway.service.LineService;
+import subway.service.SectionService;
 
 import java.net.URI;
 import java.util.List;
