+++ src/main/java/subway/controller/StationController.java
@@ -1,8 +1,12 @@
-package subway.station;
+package subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.domain.Station;
+import subway.request.StationRequest;
+import subway.response.StationResponse;
+import subway.service.StationService;
 
 import java.net.URI;
 import java.util.List;
