+++ src/main/java/subway/dao/StationDao.java
@@ -1,9 +1,10 @@
-package subway.station;
+package subway.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.domain.Station;
 
 import java.sql.PreparedStatement;
 import java.sql.Statement;
