+++ src/main/java/subway/domain/Line.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.domain;
 
 import java.util.*;
