+++ src/main/java/subway/domain/Section.java
@@ -1,6 +1,4 @@
-package subway.line;
-
-import subway.station.Station;
+package subway.domain;
 
 public class Section {
     private Long id;
