+++ src/test/java/subway/response/LineAcceptanceTest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.response;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.station.StationResponse;
+import subway.request.LineRequest;
 
 import java.util.Arrays;
 import java.util.List;
