+++ src/test/java/subway/station/StationAcceptanceTest.java
@@ -5,11 +5,11 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
+import subway.response.StationResponse;
+import subway.request.StationRequest;
 
 import java.util.Arrays;
 import java.util.List;
